Remove name duplication
